rename packages w/ proper convention, update imports
