+++ src/main/java/com/mars/trainticketbooking/springboot/TrainTicketBookingApplication.java
@@ -1,4 +1,4 @@
-package com.springboot.trainticketbooking;
+package com.mars.trainticketbooking.springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
